FIrstInFirstOut is all done, tests need to be faster

+++ src/main/kotlin/com/css/dispatcher/Statistics.kt
@@ -2,6 +2,6 @@ package com.css.dispatcher
 
 data class Statistics(
     var ordersProcessed: Int,
-    var averageFoodWaitTime: Int?,
+    var averageFoodWaitTime: Long?,
     var averageCourierWaitTime: Long?
 )
